Moved calculator test
